# docs/swagger.yaml
basePath: /api
definitions:
...
  model.DiscoveryServerResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.ServerInfoResponse'
        type: array
      page:
        $ref: '#/definitions/model.Page'
    type: object
...
  model.Page:
    properties:
      nextCursor:
        type: string
    type: object
...
  model.ServerDetailResponse:
    properties:
      avatarImageUrl:
        type: string
      bannerImageUrl:
        type: string
      categoryName:
        type: string
      createDatetime:
        type: string
      createdBy:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      shortName:
        type: string
    type: object
  model.ServerInfoForInviteResponse:
    properties:
      avatarImageId:
        type: string
      bannerImageId:
        type: string
      description:
        type: string
      ownerName:
        type: string
      serverName:
        type: string
    type: object
  model.ServerInfoResponse:
    properties:
      avatarImageUrl:
        type: string
      bannerImageUrl:
        type: string
      categoryName:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      shortName:
        type: string
    type: object
...
  model.ValidationError:
    properties:
      code:
        type: string
      message:
        type: string
      param:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: Virdan Project API Documentation
  title: Virdan API
  version: "1.0"
paths:
...
  /servers/:
    get:
      description: |
        ## Get Discovery Server Flow

        ### Overview
        Retrieves a paginated list of discoverable (public) servers. Supports cursor-based pagination and optional filtering by category.

        ### Auth
        Requires `Authorization` header with Bearer JWT access token.

        ### Business Logic
        1. Get `userId` from context
        2. Parse query parameters: `limit` (default: system default), `categoryId` (default: 0 = all), `cursor` (optional)
        3. Validate `limit` (>= 0 and <= max limit)
        4. If cursor provided, decode from base64 URL-safe encoding and unmarshal to `ServerDiscoveryCursor`
        5. Query discoverable servers from database (limit + 1 for cursor detection)
        6. Build MinIO image URLs for avatar and banner
        7. If results > limit, create next cursor from last item's ID and createDatetime
        8. Return paginated server list

        ### Query Parameters
        - `limit` — number of items per page (optional, has default)
        - `categoryId` — filter by category (optional, 0 = all categories)
        - `cursor` — base64-encoded cursor for next page (optional)

        ### Error Cases
        - Limit < 0 → `400` with "Limit must be greater or equal than 0"
        - Limit > max → `400` with "Limit is exceeded max limit: {max}"

        ### Flow
        ```
        Request → Auth Middleware → Parse Query Params → Validate Limit → Decode Cursor → Query Servers (DB) → Build Image URLs → Build Next Cursor → Response
        ```
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by category ID
        in: query
        name: categoryId
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DiscoveryServerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ValidationError'
      summary: Get discoverable servers
      tags:
      - servers
  /servers/{id}:
    delete:
      description: |
        ## Delete Server Flow

        ### Overview
        Deletes a server permanently. Only the server owner can perform this action.

        ### Auth
        Requires `Authorization` header with Bearer JWT access token.

        ### Business Logic
        1. Get `userId` from context
        2. Parse `id` from URL path — must be a valid UUID
        3. Check server ownership
        4. Delete server from database (CASCADE deletes related records)

        ### Error Cases
        - Invalid server id → `400` with "Invalid server id"
        - Not owner → `400` with "You are not the owner of this server"

        ### Flow
        ```
        Request → Auth Middleware → Parse ServerId → Check Ownership (DB) → Delete Server (DB, CASCADE) → Response (no data)
        ```
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Server UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ValidationError'
      summary: Delete a server
      tags:
      - servers
    get:
      description: |
        ## Get Server By ID Flow

        ### Overview
        Retrieves detailed information about a specific server. Access is controlled: public servers are visible to all authenticated users, private servers are only visible to members.

        ### Auth
        Requires `Authorization` header with Bearer JWT access token.

        ### Business Logic
        1. Get `userId` from context
        2. Parse `id` from URL path parameter — must be a valid UUID
        3. Query server details with a single query that checks both existence AND membership for private servers
        4. If server not found (or private and user is not a member), return error
        5. Build full MinIO URLs for avatar and banner images
        6. Return server detail

        ### Error Cases
        - Invalid server id → `400` with "Invalid server id"
        - Server not found or no permission → `400` with "Server not found or you don't have permission to access it"

        ### Flow
        ```
        Request → Auth Middleware → Parse ServerId → Query Server (DB, checks visibility + membership) → Build Image URLs → Response
        ```
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Server UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ServerDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ValidationError'
      summary: Get server by ID
      tags:
      - servers
  /servers/{id}/avatar:
    put:
      consumes:
      - multipart/form-data
      description: |
        ## Update Server Avatar Flow

        ### Overview
        Updates the avatar image of a server. Only the server owner can perform this action. Accepts multipart form file upload.

        ### Auth
        Requires `Authorization` header with Bearer JWT access token.

        ### Business Logic
        1. Get `userId` from context
        2. Parse `id` from URL path — must be a valid UUID
        3. Check server ownership
        4. Read `avatar` file from multipart form
        5. Validate image file (format, size)
        6. Begin database transaction
        7. Get current server detail to find old avatar
        8. If new avatar provided: create new image record → update server reference → upload to MinIO → delete old image if exists
        9. If no new avatar (empty file): update server to remove reference → delete old image if exists
        10. Commit transaction

        ### Error Cases
        - Invalid server id → `400` with "Invalid server id"
        - Not owner → `400` with "You are not the owner of this server"
        - Invalid image → `400` validation error

        ### Flow
        ```
        Request → Auth Middleware → Parse ServerId → Check Ownership (DB) → Read File → Validate Image → Begin TX → Handle Old/New Avatar → Commit TX → Response (no data)
        ```
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Server UUID
        in: path
        name: id
        required: true
        type: string
      - description: Avatar image file
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ValidationError'
      summary: Update server avatar
      tags:
      - servers
...
  /servers/invites/{inviteCode}:
    get:
      description: |
        ## Get Server Info For Invite Flow

        ### Overview
        Retrieves public information about a server using an invite code. This is a **public endpoint** (no authentication required) — used to show server preview before joining.

        ### Auth
        No authentication required.

        ### Business Logic
        1. Parse `inviteCode` from URL path parameter
        2. Query server info associated with the invite code
        3. If no server found for invite code, return error
        4. Build full MinIO URLs for avatar and banner images
        5. Return server info (owner name, server name, description, images)

        ### Error Cases
        - Invite code not found → `400` with "Invite code is not exists"

        ### Flow
        ```
        Request → Parse InviteCode → Query Server Info (DB) → Build Image URLs (MinIO) → Response
        ```
      parameters:
      - description: Invite code (8 characters)
        in: path
        name: inviteCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ServerInfoForInviteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ValidationError'
      summary: Get server info for invite preview
      tags:
      - server-invites
...
swagger: "2.0"